Added conditional statements for LaunchAgent setup on different platforms
Added conditional statements to set up a LaunchAgent on different platforms. For macOS, the script will check if the LaunchAgent exists and set it up if it doesn't. For Windows, for now, the script prints a message informing the user that there is no functionality to setup Task Scheduler yet and recommends creating a new task for the script.

This improves the usability of the script by providing platform-specific guidance on how to set up automated running of the script.